requirements.txt
absl-py==1.4.0
anyio==4.0.0
appdirs==1.4.4
arch==6.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
array-record==0.4.0
arrow==1.3.0
asttokens==2.4.0
astunparse==1.6.3
async-lru==2.0.4
attrs==23.1.0
Babel==2.13.1
backcall==0.2.0
backtrader==1.9.78.123
beautifulsoup4==4.12.2
bleach==6.1.0
bokeh==3.4.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
click==8.1.7
colorama==0.4.6
colorcet==3.1.0
comm==0.1.4
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
empyrical==0.5.5
etils==1.3.0
eventkit==1.0.1
exceptiongroup==1.1.3
executing==2.0.0
fastjsonschema==2.18.1
fmpsdk==20230623.0
fonttools==4.43.1
fqdn==1.5.1
frozendict==2.4.0
gast==0.4.0
google-api-core==2.11.1
google-cloud-bigquery==3.11.4
google-cloud-core==2.3.3
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.5.0
googleapis-common-protos==1.62.0
grpcio-status==1.56.2
h11==0.14.0
h5py==3.9.0
holoviews==1.18.3
html5lib==1.1
httpcore==1.0.2
httpx==0.26.0
hvplot==0.9.2
ib-insync==0.9.86
ibapi==9.81.1.post1
idna==3.4
importlib-resources==6.1.1
ipykernel==6.26.0
ipympl==0.9.3
ipython==8.16.1
ipython-genutils==0.2.0
ipywidgets==8.1.1
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.8.0
jupyter-lsp==2.2.0
jupyter_client==8.5.0
jupyter_core==5.4.0
jupyter_server==2.9.1
jupyter_server_terminals==0.4.4
jupyterlab==4.1.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
jupyterlab_server==2.25.0
kiwisolver==1.4.5
linkify-it-py==2.0.3
lxml==5.1.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdit-py-plugins==0.4.0
mdurl==0.1.2
mistune==3.0.2
multitasking==0.0.11
nbclient==0.8.0
nbconvert==7.9.2
nbformat==5.9.2
nest-asyncio==1.5.8
nltk==3.8.1
nodejs-bin==18.4.0a4
notebook==7.0.7
notebook_shim==0.2.3
numpy==1.26.1
opencv-python==4.9.0.80
opt-einsum==3.3.0
overrides==7.4.0
packaging==23.2
pandas==2.1.1
pandas-datareader==0.10.0
pandocfilters==1.5.0
panel==1.4.1
param==2.1.0
parso==0.8.3
patsy==0.5.6
peewee==3.17.0
pickleshare==0.7.5
Pillow==10.1.0
platformdirs==3.11.0
prometheus-client==0.17.1
promise==2.3
prompt-toolkit==3.0.39
proto-plus==1.22.3
psutil==5.9.6
pure-eval==0.2.2
py4j==0.10.9.7
pycparser==2.21
pyfolio==0.9.2
Pygments==2.16.1
pyparsing==3.1.1
pyspark==3.5.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-json-logger==2.0.7
pytz==2023.3.post1
pyviz_comms==3.0.2
pywin32==306
pywinpty==2.0.12
PyYAML==6.0.1
pyzmq==25.1.1
qtconsole==5.5.1
QtPy==2.4.1
referencing==0.30.2
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.10.6
scikit-learn==1.4.0
scipy==1.12.0
seaborn==0.13.0
Send2Trash==1.8.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
stack-data==0.6.3
statsmodels==0.14.1
tensorflow==2.10.1
tensorflow-datasets==4.9.4
tensorflow-metadata==1.14.0
tensorflow-text==2.10.0
terminado==0.17.1
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
torch==2.2.1
torchvision==0.17.1
tornado==6.3.3
tqdm==4.66.2
traitlets==5.12.0
types-python-dateutil==2.8.19.14
typing_extensions==4.8.0
tzdata==2023.3
uc-micro-py==1.0.3
ultralytics==8.1.32
uri-template==1.3.0
urllib3==2.0.7
wcwidth==0.2.8
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
widgetsnbextension==4.0.9
xyzservices==2023.10.1
yfinance==0.2.36
